Fix: Migrate from handlebars to ejs
