--- src/webhint-theme/layout/changelog.hbs
@@ -1,19 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>Changelog v{{page.title}} | webhint</title>
-    {{>metas}}
-
-    {{> styles}}
-</head>
-
-<body>
-    {{>gdpr}}
-    {{>header navs=site.data.menu}}
-    {{>changelog-content}}
-    {{>footer navs=site.data.menu}}
-    {{>scripts}}
-</body>
-</html>
