--- src/webhint-theme/layout/common.hbs
@@ -1,20 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>{{title}}</title>
-    {{>metas}}
-
-    {{>styles}}
-</head>
-
-<body>
-    {{>gdpr}}
-    {{>header navs=menuData}}
-    {{{body}}}
-    {{>footer navs=menuData}}
-
-    {{>scripts}}
-</body>
-</html>
