--- src/webhint-theme/layout/docs.hbs
@@ -1,54 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    {{#compare page.category '===' 'doc-index'}}
-    <title>{{page.title}} | webhint </title>
-    {{else}}
-    <title>{{page.title}} | webhint documentation</title>
-    {{/compare}}
-    {{>metas}}
-
-    {{> styles}}
-
-    <!-- build:css /static/styles/docs.css -->
-    <link rel="stylesheet" href="/core/css/code.css">
-    <link rel="stylesheet" href="/core/css/table-of-contents.css">
-    <link rel="stylesheet" href="/core/css/tables-lists.css">
-    <link rel="stylesheet" href="/components/treeview/treeview.css">
-    <link rel="stylesheet" href="/components/anchor-top/anchor-top.css">
-    <link rel="stylesheet" href="/components/placement/placement.css">
-    <link rel="stylesheet" href="/core/css/documentation.css">
-    <!-- endbuild -->
-</head>
-
-<body>
-    {{>gdpr}}
-    {{>header navs=site.data.menu}}
-    <main class="container">
-        {{#compare page.category '===' 'doc-index'}}
-            {{>doc-index}}
-        {{else}}
-        {{#compare page.category '===' 'contributor-guide'}}
-            {{>sub-page intro=site.data.developerGuide pages=(sortPagesByCategory
-            site.pages.data 'contributor-guide')}}
-        {{else}}
-        {{#compare page.category '===' 'user-guide'}}
-            {{>sub-page intro=site.data.userGuide
-            pages=(sortPagesByCategory site.pages.data 'user-guide')}}
-        {{/compare}}
-        {{/compare}}
-        {{/compare}}
-    </main>
-
-    {{>footer navs=site.data.menu}}
-
-    <!-- build:js /static/scripts/treeview.js -->
-    <script async src="/js/details-polyfill.js"></script>
-    <script async src="/components/treeview/treeview.js"></script>
-    <script async src="/components/anchor-top/anchor-top.js"></script>
-    <!-- endbuild -->
-    {{>scripts}}
-</body>
-</html>
