--- src/webhint-theme/layout/empty.hbs
@@ -1,1 +1,0 @@
-{{{body}}}
