--- src/webhint-theme/layout/index.hbs
@@ -1,24 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>webhint, the hinting engine for web best practices</title>
-    {{>metas}}
-
-    {{>styles}}
-    <!-- build:css /static/styles/home.css -->
-    <link rel="stylesheet" href="/core/css/home.css">
-    <!-- endbuild -->
-</head>
-
-<body>
-    {{>gdpr}}
-
-    {{>header navs=site.data.menu}}
-    {{>home}}
-    {{>footer navs=site.data.menu}}
-
-    {{>scripts}}
-</body>
-</html>
