+++ src/webhint-theme/layout/partials/browser-icon.ejs
@@ -1,0 +1,11 @@
+<% if (result && result.status) { %>
+    <% if (result.status === 'error' || result.errors !== 0 || result.warnings !== 0) { %>
+        <link rel="icon" href="/images/favicon_failed.ico">
+    <% } else if (result.status === 'finished') { %>
+        <link rel="icon" href="/images/favicon_passed.ico">
+    <% } else { %>
+        <link rel="icon" href="/images/favicon_pending.ico">
+    <% } %>
+<% } else { %>
+    <link rel="icon" href="/images/favicon.ico">
+<% } %>
