+++ src/webhint-theme/layout/partials/doc-index.ejs
@@ -8,7 +8,7 @@
         </nav>
         <div class="section">
             <div class="documentation-wrapper">
-                {{>search-form}}
+                <%- include('search-form'); %>
                 <div class="documentation">
                     <p class="subtitle">Or browse the documentation by guide</p>
                     <ul class="bare">
