+++ src/webhint-theme/layout/partials/error.ejs
@@ -4,8 +4,8 @@
             <img src="/images/404-space-narwhal.svg" alt="404 narhwal lost in space">
         </div>
         <div class="module--secondary">
-            <h1 id="page-heading">{{ heading }}</h1>
-            <h2>{{ details }}</h2>
+            <h1 id="page-heading"><%= heading %></h1>
+            <h2><%= details %></h2>
         </div>
     </div>
 </section>
