+++ src/webhint-theme/layout/partials/footer.ejs
@@ -8,7 +8,7 @@
             <ul>
                 <% nav.items && nav.items.forEach((item) => { %>
                 <li>
-                    <a href="<%= item.link %>"><%= item.title %></a>
+                    <a href="<%= item.link %>" data-ga-category="docs" data-ga-action="<%= item.title %>" data-ga-label="footer"><%= item.title %></a>
                 </li>
                 <% }); %>
             </ul>
