--- src/webhint-theme/layout/partials/footer.hbs
@@ -1,35 +1,0 @@
-<footer class="footer footer--home">
-    <div class="footer-nav container">
-        {{#each navs}}
-        <div class="footer-nav__item">
-            <p class="category">
-                <a href="{{link}}">{{title}}</a>
-            </p>
-            <ul>
-                {{#each items}}
-                <li>
-                    <a href="{{link}}" data-ga-category="docs" data-ga-action="{{title}}" data-ga-label="footer" >{{title}}</a>
-                </li>
-                {{/each}}
-            </ul>
-        </div>
-        {{/each}}
-        <ul class="footer-nav__item social-list">
-            <li class="github-icon">
-                <a href="https://github.com/webhintio" aria-label="Follow webhint on GitHub">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 25 25"><path class="github-icon-svg" d="M12.5.4C5.7.4.2 5.9.2 12.7c0 5.4 3.5 10 8.4 11.7.6.1.8-.3.8-.6v-2.1C6 22.4 5.3 20 5.3 20c-.6-1.4-1.4-1.8-1.4-1.8-1.1-.7.1-.7.1-.7 1.2.1 1.9 1.3 1.9 1.3 1.1 1.9 2.9 1.3 3.6 1 .1-.8.4-1.3.8-1.6-2.7-.3-5.6-1.4-5.6-6.1 0-1.3.5-2.4 1.3-3.3-.2-.4-.6-1.6 0-3.3 0 0 1-.3 3.4 1.3 1-.3 2-.4 3.1-.4 1 0 2.1.1 3.1.4C17.9 5.2 19 5.5 19 5.5c.7 1.7.2 2.9.1 3.3.8.9 1.3 2 1.3 3.3 0 4.7-2.9 5.8-5.6 6.1.4.4.8 1.1.8 2.3v3.4c0 .3.2.7.8.6 4.9-1.6 8.4-6.2 8.4-11.7C24.8 5.9 19.3.4 12.5.4"/></svg>
-                </a>
-            </li>
-            <li class="twitter-icon">
-                <a href="https://twitter.com/narwhalnellie" aria-label="Follow webhint on Twitter">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 25 24"><path class="twitter-icon-svg" d="M25 4.4c-.9.4-1.9.7-2.9.8 1-.6 1.8-1.6 2.2-2.8-1 .6-2.1 1-3.2 1.2-.9-1-2.2-1.6-3.7-1.6-2.8 0-5 2.3-5 5 0 .4 0 .8.1 1.2C8.3 8 4.6 6 2.1 2.9c-.4.8-.7 1.6-.7 2.6 0 1.8.9 3.3 2.2 4.2-.8-.1-1.5-.3-2.2-.7v.1c0 2.4 1.7 4.5 4.1 5-.4.1-.9.2-1.3.2-.3 0-.6 0-.9-.1.6 2 2.5 3.5 4.7 3.5-1.9 1.3-4 2.1-6.4 2.1-.4 0-.8 0-1.2-.1C2.6 21.2 5.3 22 8.1 22c9.3 0 14.4-7.7 14.4-14.4v-.7c1-.6 1.8-1.5 2.5-2.5"/></svg>
-                </a>
-            </li>
-            <li class="medium-icon">
-                <a href="https://medium.com/webhint" aria-label="Follow webhint on Medium">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="4 2 40 40"><path d="M5 40V5h35v35H5zm8.56-12.63c0 .56-.03.7-.32 1.03l-2.45 3v.38h6.9v-.4l-2.4-2.98c-.3-.34-.35-.5-.35-1.03v-8.95l6.13 13.36h.7L27 18.42v10.64c0 .3 0 .35-.2.53L25 31.4v.37h9.2v-.4l-1.82-1.8c-.2-.17-.22-.23-.22-.52v-13.1c0-.3.03-.36.2-.54l1.83-1.8v-.32h-6.5L23.1 24.8l-5.2-11.55h-6.8v.4l2.14 2.63c.24.3.3.37.3.77V27.4z" class="medium-icon-svg"/></svg>
-                </a>
-            </li>
-        </ul>
-    </div>
-</footer>
