--- src/webhint-theme/layout/partials/gdpr.hbs
@@ -1,4 +1,0 @@
-<div id="disclaimer" class="disclaimer">
-    This site uses cookies for analytics. By continuing to browse this site, you agree to this use.
-    <button class="close" aria-label="close">X</button>
-</div>
