+++ src/webhint-theme/layout/partials/header.ejs
@@ -23,7 +23,7 @@
     </div>
     <div class="nav-bar container">
         <nav class="nav">
-            <a href="/" class="clear logo--container"><img src="/images/sonar-logo.svg" alt="webhint logo" class="header__logo navbar__navitem"></a>
+        <a href="/" class="clear logo--container"><img src="/images/webhint-logo.svg" alt="webhint" class="header__logo navbar__navitem"></a>
             <ul class="nav__navbar">
                 <% navs.forEach((nav) => { %>
                     <li class="navbar__navitem">
