+++ src/webhint-theme/layout/partials/home.ejs
