--- src/webhint-theme/layout/partials/metas.hbs
@@ -1,27 +1,0 @@
-<meta property="og:image" content="{{setDefault page.image '/images/apple-touch-icon.png'}}" />
-<meta property="og:description" content="{{setDefault page.description 'webhint, a linting tool for the web focused on best practices and flexibility for the end user'}}" />
-<meta property="og:title" content="{{page.title}}" />
-
-<meta name="description" content="{{setDefault page.description 'webhint, a linting tool for the web focused on best practices and flexibility for the end user'}}">
-<meta name="viewport" content="width=device-width">
-
-<meta name="theme-color" content="#4046dd">
-<link rel="manifest" href="/static/site.webmanifest">
-
-<link rel="apple-touch-icon" href="/images/apple-touch-icon.png">
-<link type="application/opensearchdescription+xml" rel="search" href="/static/opensearch.xml" title="webhint">
-
-{{#if status}}
-    {{#if (isFinish status)}}
-        {{#if (isPass status overallStatistics)}}
-            <link rel="icon" href="/images/favicon_passed.ico">
-        {{else}}
-            <link rel="icon" href="/images/favicon_failed.ico">
-        {{/if}}
-    {{else}}
-        <link rel="icon" href="/images/favicon_pending.ico">
-    {{/if}}
-{{else}}
-    <link rel="icon" href="/images/favicon.ico">
-{{/if}}
-
