--- src/webhint-theme/layout/partials/pagination.hbs
@@ -1,34 +1,0 @@
-{{#pagination currentPage totalPages pattern}}
-    {{#unless hidePagination}}
-    <ul>
-        {{#if prev}}
-        <li class="pagination__prev">
-            <a rel="prev" href="{{prev.url}}" aria-label="Previous page">
-                <span>Previous</span>
-            </a>
-        </li>
-        {{/if}}
-        {{#each pages}}
-            {{#each this}}
-                <li>
-                    {{#if this.active}}
-                    <span aria-current="page" aria-label="Current page: {{this.page}}">{{this.page}}</span>
-                    {{else}}
-                    <a href="{{url}}" aria-label="Page {{this.page}}">{{this.page}}</a>
-                    {{/if}}
-                </li>
-            {{/each}}
-            {{#unless @last}}
-                <li>...</li>
-            {{/unless}}
-        {{/each}}
-        {{#if next}}
-        <li class="pagination__next">
-            <a rel="next" href="{{next.url}}" aria-label="Next page">
-                <span>Next</span>
-            </a>
-        </li>
-        {{/if}}
-    </ul>
-    {{/unless}}
-{{/pagination}}
