--- src/webhint-theme/layout/partials/rule-category.hbs
@@ -1,12 +1,0 @@
-<div class="subtitle category-description">
-    <p class="subtitle">{{description}}</p>
-</div>
-<div class="layout">
-    <div class="module">
-       <ul>
-            {{#each hints}}
-            <li><a href="{{link}}"><code>{{name}}</code></a></li>
-            {{/each}}
-        </ul>
-    </div>
-</div>
