+++ src/webhint-theme/layout/partials/rules-index.ejs
@@ -1,0 +1,12 @@
+<div class="layout layout--halves">
+    <% categories.forEach((category) => { %>
+    <div class="module placement">
+        <p class="title icon--categories icon--<%= category.normalizedName %>"><%= category.name %></p>
+        <p class="placement__body"><%= category.description %></p>
+        <div class="placement__rules">
+            <p class="subtitle--alt"><%= utils.getHintsCount(category.hints);%> Hints</p>
+            <a class="button--alt placement__button" href="<%= category.categorylink %>">View All</a>
+        </div>
+    </div>
+    <% }); %>
+</div>
