--- src/webhint-theme/layout/partials/rules-index.hbs
@@ -1,12 +1,0 @@
-<div class="layout layout--halves">
-    {{#each categories}}
-    <div class="module placement">
-        <p class="title icon--categories icon--{{normalizedName}}">{{name}}</p>
-        <p class="placement__body">{{description}}</p>
-        <div class="placement__rules">
-            <p class="subtitle--alt">{{getHintsCount hints}} Hints</p>
-            <a class="button--alt placement__button" href="{{link}}">View All</a>
-        </div>
-    </div>
-    {{/each}}
-</div>
