--- src/webhint-theme/layout/partials/scan-error-message.hbs
@@ -1,6 +1,0 @@
-<div class="scan-error">
-    <p>
-        There was an error and we were only able to partially complete the scan. View the results below or
-        <a href="/scanner/">perform another scan</a>.
-    <p>
-</div>
