+++ src/webhint-theme/layout/partials/scan-error.ejs
