--- src/webhint-theme/layout/partials/scan.ejs
@@ -1,50 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title><%= page.title %></title>
-    <%- include('metas', { page, result }); %>
-
-    <%- include('styles'); %>
-    <!-- build:css /static/styles/scanner.css -->
-    <link rel="stylesheet" href="/core/css/scanner.css">
-    <link rel="stylesheet" href="/core/css/scan/scan-results.css">
-    <!-- endbuild -->
-    <!-- build:css /static/styles/highlight.css -->
-    <link href="/components/highlight/default.min.css" rel="stylesheet">
-    <!-- endbuild -->
-</head>
-
-<body>
-    <%- include('gdpr') %>
-
-    <%- include('header', { navs: menuData }); %>
-    
-    <% if (result) { %>
-        <%- include('scan-result-container', { result, utils, showQueue }); %>
-    <% } else { %>
-        <%- include('scan-form'); %>
-    <% } %>
-
-    <%- include('footer', { navs: menuData }); %>
-
-    <%- include('scripts', { theme }); %>
-    <!-- build:js /static/scripts/highlight.js -->
-    <script src="/js/highlight/highlight.min.js"></script>
-    <script src="/js/highlight/languages/xml.min.js"></script>
-    <!-- endbuild -->
-    <!-- build:js /static/scripts/partials.js -->
-    <script src="/js/partials/scan-result-item.js"></script>
-    <script src="/js/partials/category-pass-message.js"></script>
-    <script src="/js/partials/scan-error-message.js"></script>
-    <script src="/js/partials/utils.js"></script>
-    <!-- endbuild -->
-    <% if(!result || !result.isFinish) { %>
-    <!-- build:js /static/scripts/scanner.js -->
-        <script src="/js/scan/scanner-common.js"></script>
-        <script src="/js/scanner-submit.js"></script>
-    <!-- endbuild -->
-    <% } %>
-</body>
-</html>
