--- src/webhint-theme/layout/partials/styles.hbs
@@ -1,20 +1,0 @@
-<!-- build:css /static/styles/common.css -->
-<link rel="stylesheet" href="/core/css/fonts.css">
-<link rel="stylesheet" href="/core/css/about.css">
-<link rel="stylesheet" href="/core/css/base.css">
-<link rel="stylesheet" href="/core/css/color.css">
-<link rel="stylesheet" href="/core/css/footer.css">
-<link rel="stylesheet" href="/core/css/helpers.css">
-<link rel="stylesheet" href="/core/css/layouts.css">
-<link rel="stylesheet" href="/core/css/media.css">
-<link rel="stylesheet" href="/core/css/navigation.css">
-<link rel="stylesheet" href="/core/css/search-results.css">
-<link rel="stylesheet" href="/core/css/structure.css">
-<link rel="stylesheet" href="/core/css/type.css">
-<link rel="stylesheet" href="/core/css/highlight.css">
-<link rel="stylesheet" href="/components/breadcrumb/breadcrumb.css">
-<link rel="stylesheet" href="/components/pagination/pagination.css">
-<link rel="stylesheet" href="/core/css/controls.css">
-<link rel="stylesheet" href="/core/css/scanner.css">
-<!-- endbuild -->
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/docsearch.js/2/docsearch.min.css" integrity="sha384-mHx5KeR3zcLviQFPmpHqrwf2uxtukRvUFVzOMOqIjKjSNrurxlmJ3oS+vm49Mz07" crossorigin="anonymous">
