--- src/webhint-theme/layout/partials/sub-page.hbs
@@ -1,118 +1,0 @@
-<article class="content" aria-labelledby="page-heading">
-    <nav class="breadcrumb" aria-label="breadcrumbs">
-        <ul>
-            <li>
-                <a href="/">Home</a>
-            </li>
-            <li>
-                <a href="{{getTargetPagePath page 'docs'}}">Documentation</a>
-            </li>
-            {{#if (isNotGuideIndexPage page)}}
-            <li>
-                <a href="{{getTargetPagePath page 'category'}}" class="guide-category">
-                    {{capitalize page.category}}
-                </a>
-            </li>
-            {{/if}}
-            <li aria-current="location">{{page.title}}</li>
-        </ul>
-    </nav>
-    <header class="page-intro">
-        <h1 id="page-heading">{{page.title}}</h1>
-        <div class="documentation-actions">
-            <ul>
-                <li>
-                    <a target="_blank" href="{{theme.webhintRepoRoot}}/edit/master/{{getEditLink page.originalFile}}">Edit</a>
-
-                </li>
-                <li>
-                    <a target="_blank" href="{{theme.webhintRepoRoot}}/issues/new?{{getSignalIssueQuery theme.webhintRoot page.title page.permalink}}">Signal Issue</a>
-                </li>
-            </ul>
-        </div>
-    </header>
-    <div class="section" id="main">
-        <div class="layout layout--sidebar">
-            <div class="module module--secondary mobile-view">
-                <ul class="toc-mobile-view">
-                    <li>
-                        <p role="button" id="mobile-toc" aria-controls="mobile-toc-submenu" aria-expanded="false" aria-haspopup="true" class="toc-mobile-view__button expandable">{{capitalize page.category}}</p>
-                        <ul role="group" class="toc-mobile-view__dropdown" id="mobile-toc-submenu" aria-hidden="true">
-                            {{#if (isNotGuideIndexPage page)}}
-                            <li class="dropdown__subitem">
-                                <a href="../{{#if (isNotSectionIndexPage page)}}../{{/if}}" class="subitem--single">Getting Started</a>
-                            </li>
-                            {{else}}
-                            <li class="dropdown__subitem">
-                                <a href="#" class="subitem--single">Getting Started</a>
-                            </li>
-                            {{/if}} {{#each (getSortedToCTitles pages)}}
-                            <li class="dropdown__subitem">
-                                <p>{{capitalize this}}</p>
-                                <ul>
-                                    {{#each (getPagesByToCTitle this ../pages)}}
-                                    <li class="subitem">
-                                        <a href="{{sanitize permalink}}">{{title}}</a>
-                                    </li>
-                                    {{/each}}
-                                </ul>
-                            </li>
-                            {{/each}}
-                        </ul>
-                    </li>
-                </ul>
-            </div>
-            <div class="module module--secondary table-of-contents treeview main-view" role="navigation">
-                <ul role="tree" aria-labelledby="page-heading" tabindex="0">
-                    {{#each (getSortedToCTitles pages)}}
-                    <li role="treeitem" aria-label="subsection" tabindex="0" aria-expanded="{{#compare ../page.tocTitle '===' this}}true{{else}}false{{/compare}}">
-                        <div class="
-                        {{#if (isActiveItem ../page this)}}
-                            toc-section-title--active
-                        {{else}}
-                            toc-section-title
-                        {{/if}}">
-                            <span class="text">{{capitalize this}}</span>
-                            <span class="icon"></span>
-                        </div>
-
-                        <ul role="group">
-                            {{#each (getPagesByToCTitle this ../pages)}}
-                            <li role="treeitem" aria-selected="{{#compare ../../page.title '===' title}}true{{else}}false{{/compare}}">
-                                <div class="
-                                {{#compare ../../page.title '===' title}}
-                                    toc-subsection-title--active
-                                {{else}}
-                                    toc-subsection-title
-                                {{/compare}}"
-                                    aria-current="{{#compare ../../page.title '===' title}}location{{/compare}}">
-                                    <a class="text" href="{{sanitize permalink}}" tabindex="{{#compare ../../page.tocTitle '!==' ../this}}-1{{else}}0{{/compare}}">
-                                        {{title}}
-                                    </a>
-                                </div>
-                            </li>
-                            {{/each}}
-                        </ul>
-                    </li>
-                    {{/each}}
-                </ul>
-            </div>
-            <div class="module module--primary wrap-text">
-                {{#if (showMdContent page)}}
-                    {{{page.content}}}
-                {{else}}
-                    {{#compare page.title '===' 'Hints categories'}}
-                        {{>rules-index site.data.categories}}
-                    {{else}}
-                        {{#each site.data.categories.categories}}
-                            {{#compare name '===' ../page.title}}
-                                {{>rule-category}}
-                            {{/compare}}
-                        {{/each}}
-                    {{/compare}}
-                {{/if}}
-            </div>
-        </div>
-        <a href="#main" class="anchor-top" hidden aria-label="back to top">top</a>
-    </div>
-</article>
