+++ src/webhint-theme/layout/partials/under-construction.ejs
@@ -4,7 +4,7 @@
             <img src="/images/nellie-construction.svg" alt="Website is currently under construction">
         </div>
         <div class="module--secondary">
-            <h1 id="page-heading">{{ heading }}</h1>
+            <h1 id="page-heading"><%= heading %></h1>
             <h2>Please excuse Nellie's interruption: her and her team are currently working on improving the service. Please check back soon.</h2>
         </div>
     </div>
