+++ src/webhint-theme/layout/scan.ejs
@@ -1,0 +1,60 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <title><%= page.title %></title>
+    <%- include('partials/metas', { page }); %>
+    <%- include('partials/browser-icon', {result}); %>
+
+    <%- include('partials/styles'); %>
+    <!-- build:css /static/styles/scanner.css -->
+    <link rel="stylesheet" href="/core/css/scanner.css">
+    <link rel="stylesheet" href="/core/css/scan/scan-results.css">
+    <!-- endbuild -->
+    <!-- build:css /static/styles/highlight.css -->
+    <link href="/components/highlight/default.min.css" rel="stylesheet">
+    <!-- endbuild -->
+</head>
+
+<body>
+    <%- include('partials/gdpr') %>
+
+    <%- include('partials/header', { navs: menuData }); %>
+    
+    <% if (result) { %>
+        <%- include('partials/scan-result-container', { result, utils, showQueue }); %>
+    <% } else { %>
+        <%- include('partials/scan-form'); %>
+    <% } %>
+
+    <%- include('partials/footer', { navs: menuData }); %>
+
+    <%- include('partials/scripts', { theme }); %>
+
+    <% if(result) { %>
+        <!-- build:js /static/scripts/highlight.js -->
+        <script src="/js/highlight/highlight.min.js"></script>
+        <script src="/js/highlight/languages/xml.min.js"></script>
+        <!-- endbuild -->
+
+        <!-- build:js /static/scripts/partials.js -->
+        <script src="/js/partials/scan-result-item.js"></script>
+        <script src="/js/partials/category-pass-message.js"></script>
+        <script src="/js/partials/scan-error-message.js"></script>
+        <script src="/js/partials/utils.js"></script>
+        <!-- endbuild -->
+
+        <% if (result.isFinish) { %>
+            <!-- build:js /static/scripts/scanner-finish.js -->
+            <script src="/js/scan/scanner-common.js"></script>
+            <!-- endbuild -->
+        <% } else { %>
+        <!-- build:js /static/scripts/scanner.js -->
+        <script src="/js/scan/scanner-common.js"></script>
+        <script src="/js/scanner-submit.js"></script>
+        <!-- endbuild -->
+        <% } %>
+    <% } %>
+</body>
+</html>
