--- src/webhint-theme/layout/scan.hbs
@@ -1,52 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>{{page.title}}</title>
-    {{>metas}}
-
-    {{>styles}}
-    <!-- build:css /static/styles/scanner.css -->
-    <link rel="stylesheet" href="/core/css/scanner.css">
-    <link rel="stylesheet" href="/core/css/scan-results.css">
-    <!-- endbuild -->
-    <!-- build:css /static/styles/highlight.css -->
-    <link href="/components/highlight/default.min.css" rel="stylesheet">
-    <!-- endbuild -->
-</head>
-
-<body>
-    {{>gdpr}}
-
-    {{>header navs=menuData}}
-    {{{body}}}
-    {{>footer navs=menuData}}
-
-    {{>scripts}}
-    <!-- build:js /static/scripts/handlebars.js -->
-    <script src="/js/handlebars/handlebars.min.js"></script>
-    <!-- endbuild -->
-    <!-- build:js /static/scripts/highlight.js -->
-    <script src="/js/highlight/highlight.min.js"></script>
-    <script src="/js/highlight/languages/xml.min.js"></script>
-    <!-- endbuild -->
-    <!-- build:js /static/scripts/helpers.js -->
-    <script src="/js/helpers/compare.js"></script>
-    <script src="/js/helpers/cutstring.js"></script>
-    <script src="/js/helpers/getlength.js"></script>
-    <script src="/js/helpers/linkify.js"></script>
-    <script src="/js/helpers/normalizeposition.js"></script>
-    <script src="/js/helpers/compare.js"></script>
-    <!-- endbuild -->
-    <!-- build:js /static/scripts/partials.js -->
-    <script src="/js/partials/scan-result-item.js"></script>
-    <script src="/js/partials/category-pass-message.js"></script>
-    <script src="/js/partials/scan-error-message.js"></script>
-    <!-- endbuild -->
-    <!-- build:js /static/scripts/scanner.js -->
-    <script src="/js/scanner-common.js"></script>
-    <script src="/js/scanner-submit.js"></script>
-    <!-- endbuild -->
-</body>
-</html>
